Introduce 'isTaskAdded()' method
With this method it is possible to check if tasks are already added
to the task chain of the scheduler. This may be used in situations
where adding/deleting the same task twice could happen frequently
(which causes crashes or undetermined results with this scheduler).
Checking the result with this method before the task add/delete
operation prevents this behaviour.